#3788 Moved test to module tests.test_model_serializer
